Ensure best available precision for floating point result.

+++ src/inline-fn.h
@@ -192,7 +192,7 @@ static inline uint32_t ballot_max3(uint32_t a, uint32_t b, uint32_t c)
 
 static inline double timeval_to_float(struct timeval tv)
 {
-	return tv.tv_sec + tv.tv_usec*1e-6;
+	return tv.tv_sec + tv.tv_usec*(double)1.0e-6;
 }
 
 static inline int timeval_msec(struct timeval tv)
